Move the ToDescriptiveString method to the generic type. Validates that the givens are correct.
