+++ Jinaga/Projections/Specification.cs
@@ -60,32 +60,7 @@ namespace Jinaga.Projections
             return $"{indent}({given}) {{\n{matches}{indent}}}{projection}\n";
         }
 
-        public string ToDescriptiveString<TFact>(TFact given) where TFact: class
-        {
-            var collector = new Collector(SerializerCache.Empty);
-            var givenReference = collector.Serialize(given);
-            var givenReferences = ImmutableList.Create(givenReference);
-
-            string startString = GenerateDeclarationString(givenReferences);
-            string specificationString = ToDescriptiveString();
-            return startString + "\n" + specificationString;
-        }
-
-        public string ToDescriptiveString<TFactA, TFactB>(TFactA a, TFactB b)
-            where TFactA : class
-            where TFactB : class
-        {
-            var collector = new Collector(SerializerCache.Empty);
-            var aReference = collector.Serialize(a);
-            var bReference = collector.Serialize(b);
-            var givenReferences = ImmutableList.Create(aReference, bReference);
-
-            string startString = GenerateDeclarationString(givenReferences);
-            string specificationString = ToDescriptiveString();
-            return startString + "\n" + specificationString;
-        }
-
-        private string GenerateDeclarationString(ImmutableList<Facts.FactReference> givenReferences)
+        protected string GenerateDeclarationString(ImmutableList<Facts.FactReference> givenReferences)
         {
             var startStrings = Given.Zip(givenReferences, (label, reference) =>
                 $"let {label.Name}: {reference.Type} = #{reference.Hash}\n");
