+++ Jinaga/Repository/FactRepository.cs
@@ -7,11 +7,14 @@ namespace Jinaga.Repository
 {
     public abstract class FactRepository
     {
-        public abstract IQueryable<TFact> OfType<TFact>();
-        public abstract IQueryable<TFact> OfType<TFact>(Expression<Func<TFact, bool>> predicate);
+        public abstract IQueryable<TFact> OfType<TFact>()
+            where TFact : class;
+        public abstract IQueryable<TFact> OfType<TFact>(Expression<Func<TFact, bool>> predicate)
+            where TFact : class;
         public abstract IObservableCollection<TProjection> Observable<TFact, TProjection>(
             TFact start,
-            Specification<TFact, TProjection> specification);
+            Specification<TFact, TProjection> specification)
+            where TFact : class;
         public abstract IObservableCollection<TProjection> Observable<TProjection>(IQueryable<TProjection> queryable);
     }
 }
