+++ Jinaga/Specification.cs
@@ -1,6 +1,7 @@
 using Jinaga.Pipelines;
 using Jinaga.Projections;
 using Jinaga.Repository;
+using Jinaga.Serialization;
 using System;
 using System.Collections.Immutable;
 using System.Linq;
@@ -10,6 +11,7 @@ namespace Jinaga
 {
 
     public static class Given<TFact>
+        where TFact: class
     {
         public static Specification<TFact, TProjection> Match<TProjection>(Expression<Func<TFact, FactRepository, IQueryable<TProjection>>> specExpression)
         {
@@ -25,6 +27,8 @@ namespace Jinaga
     }
 
     public static class Given<TFact1, TFact2>
+        where TFact1 : class
+        where TFact2 : class
     {
         public static Specification<TFact1, TFact2, TProjection> Match<TProjection>(Expression<Func<TFact1, TFact2, FactRepository, IQueryable<TProjection>>> specExpression)
         {
@@ -40,18 +44,43 @@ namespace Jinaga
     }
 
     public class Specification<TFact, TProjection> : Specification
+        where TFact : class
     {
         public Specification(ImmutableList<Label> given, ImmutableList<Match> matches, Projection projection)
             : base(given, matches, projection)
         {
         }
+
+        public string ToDescriptiveString(TFact given)
+        {
+            var collector = new Collector(SerializerCache.Empty);
+            var givenReference = collector.Serialize(given);
+            var givenReferences = ImmutableList.Create(givenReference);
+
+            string startString = GenerateDeclarationString(givenReferences);
+            string specificationString = ToDescriptiveString();
+            return startString + "\n" + specificationString;
+        }
     }
 
     public class Specification<TFact1, TFact2, TProjection> : Specification
+        where TFact1: class
+        where TFact2 : class
     {
         public Specification(ImmutableList<Label> given, ImmutableList<Match> matches, Projection projection)
             : base(given, matches, projection)
         {
+        }
+        public string ToDescriptiveString(TFact1 given1, TFact2 given2)
+        {
+            var collector = new Collector(SerializerCache.Empty);
+            var aReference = collector.Serialize(given1);
+            var bReference = collector.Serialize(given2);
+            var givenReferences = ImmutableList.Create(aReference, bReference);
+
+            string startString = GenerateDeclarationString(givenReferences);
+            string specificationString = ToDescriptiveString();
+            return startString + "\n" + specificationString;
         }
     }
 }
